updated version providing better performance on x86-64
git-svn: trunk@2559

+++ clamav-devel/libclamav/md5.h
@@ -1,16 +1,18 @@
 /*
  * This is an OpenSSL-compatible implementation of the RSA Data Security,
- * Inc. MD5 Message-Digest Algorithm.
+ * Inc. MD5 Message-Digest Algorithm (RFC 1321).
  *
  * Written by Solar Designer <solar at openwall.com> in 2001, and placed
- * in the public domain.  See md5.c for more information.
+ * in the public domain.  There's absolutely no warranty.
+ *
+ * See md5.c for more information.
  */
 
 #ifndef __MD5_H
 #define __MD5_H
 
 /* Any 32-bit or wider unsigned integer data type will do */
-typedef unsigned long MD5_u32plus;
+typedef unsigned int MD5_u32plus;
 
 typedef struct {
 	MD5_u32plus lo, hi;
